refactor(protocol): sub-split control-room schema into per-tab files (#6272)
The 1350-line schemas/server/control-room.ts is now a thin sub-barrel that
re-exports eight per-tab slices under ./control-room/ (host, runners,
containers, byok, integrations, host-prune, device-runtimes, skills). Pure
move: every export and schema body is the verbatim original; the only added
edge is integrations.ts importing ByokPoolLimitsSchema from ./byok.ts for the
pool-resize ack. ../server.ts and every downstream consumer import is
unchanged. Follow-up to #6271 (Tier-3 server.ts barrel split).

Rebuilds + force-adds the committed dist/ slices (.js + .d.ts, matching the
#6271 convention) so the no-build-step server keeps resolving @chroxy/protocol.

Makes store-core's protocol-type-coverage-lint read the schemas/server/ tree
recursively: the ~22 control-room z.literal discriminants now live one level
down, and the prior non-recursive read would have silently emptied the
control-room surface (tsc stays green; only the lint catches it). Type count
stays in the [80,120] band.
